+++ src/cache_type.d.ts
@@ -1,0 +1,4 @@
+export interface cache_type<I extends unknown = unknown> {
+    data: I;
+    promise_a1: Promise<I>[];
+}