+++ src/index.d.ts
@@ -1,3 +1,3 @@
 export * from './_ensure_store_cache';
 export * from './_reload_store_cache';
-export * from './cache_store_type';
+export * from './cache_type';