+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ctx-core/cache
 
+## 6.0.1
+
+### Patch Changes
+
+- \_reload_store_cache,\_ensure_store_cache: S generic which is a Writable<cache_store_type<I>>
+- $cache_store_type->cache_type
+
 ## 6.0.0
 
 ### Major Changes